+++ ex/app/components/guide/guide.module.ts
@@ -3,7 +3,7 @@ import { NgModule } from '@angular/core'
 import { RouterModule } from '@angular/router'
 import { FormsModule } from '@angular/forms'
 import { GuideRoutingModule } from './guide.routing'
-import { ElModule } from '../../../../src/index'
+import { ElModule } from '../../../../src/element-angular.module'
 import { ExSharedModule } from '../../shared/module'
 import { ExGuideMainComponent } from './main/main.component'
 import { ExInstallComponent } from './install/install.component'
