+++ ex/app/components/module.ts
@@ -2,7 +2,7 @@ import { CommonModule } from '@angular/common'
 import { NgModule } from '@angular/core'
 import { FormsModule } from '@angular/forms'
 import { RouterModule } from '@angular/router'
-import { ElModule } from '../../../src/index'
+import { ElModule } from '../../../src/element-angular.module'
 
 import { ExHeaderComponent } from './header/header.component'
 import { ExSideComponent } from './side/side.component'
