+++ ex/app/notice/message/message.component.ts
@@ -1,5 +1,5 @@
 import { Component, forwardRef, Inject, ViewEncapsulation } from '@angular/core'
-import { ElMessageService } from '../../../../src/index'
+import { ElMessageService } from '../../../../src/element-angular.module'
 import code from './code'
 
 @Component({
