+++ ex/app/shared/components/demo/demo.component.ts
@@ -3,7 +3,7 @@ import { trigger, state, style, animate, transition } from '@angular/animations'
 import { SafeHtml, DomSanitizer } from '@angular/platform-browser'
 import { ActivatedRoute } from '@angular/router'
 import { HighLightPipe } from '../../pipe'
-import { ElMessageService } from '../../../../../src/index'
+import { ElMessageService } from '../../../../../src/element-angular.module'
 
 @Component({
   selector: 'ex-demo',
